Rudimentary implementation of Import
