+++ src/lib/leftSidebar/Create.svelte
@@ -1,0 +1,2 @@
+Create your own scenario by changing values on the map. Coming soon to a web app
+near you...
